Load lambda types in the application layer by reusing lambda serialization code
