+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/heap/ImageLayerSnapshotUtil.java
@@ -102,6 +102,8 @@ public class ImageLayerSnapshotUtil {
     public static final String INTERFACES_TAG = "interfaces";
     public static final String WRAPPED_TYPE_TAG = "wrapped type";
     public static final String GENERATED_SERIALIZATION_TAG = "generated serialization";
+    public static final String LAMBDA_TYPE_TAG = "lambda type";
+    public static final String CAPTURING_CLASS_TAG = "capturing class";
     public static final String RAW_DECLARING_CLASS_TAG = "raw declaring class";
     public static final String RAW_TARGET_CONSTRUCTOR_CLASS_TAG = "raw target constructor class";
     public static final String CONSTANTS_TAG = "constants";
