+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/bootstrap/BootstrapMethodConfiguration.java
@@ -69,6 +69,8 @@ public class BootstrapMethodConfiguration implements InternalFeature {
     private final ConcurrentMap<BootstrapMethodRecord, BootstrapMethodInfo> bootstrapMethodInfoCache = new ConcurrentHashMap<>();
     private final Set<Executable> indyBuildTimeAllowList;
     private final Set<Executable> condyBuildTimeAllowList;
+    private final Method metafactory;
+    private final Method altMetafactory;
 
     public static BootstrapMethodConfiguration singleton() {
         return ImageSingletons.lookup(BootstrapMethodConfiguration.class);
@@ -79,10 +81,10 @@ public class BootstrapMethodConfiguration implements InternalFeature {
          * Bootstrap method used for Lambdas. Executing this method at run time implies defining
          * hidden class at run time, which is unsupported.
          */
-        Method metafactory = ReflectionUtil.lookupMethod(LambdaMetafactory.class, "metafactory", MethodHandles.Lookup.class, String.class, MethodType.class, MethodType.class, MethodHandle.class,
+        metafactory = ReflectionUtil.lookupMethod(LambdaMetafactory.class, "metafactory", MethodHandles.Lookup.class, String.class, MethodType.class, MethodType.class, MethodHandle.class,
                         MethodType.class);
         /* Alternate version of LambdaMetafactory.metafactory. */
-        Method altMetafactory = ReflectionUtil.lookupMethod(LambdaMetafactory.class, "altMetafactory", MethodHandles.Lookup.class, String.class, MethodType.class, Object[].class);
+        altMetafactory = ReflectionUtil.lookupMethod(LambdaMetafactory.class, "altMetafactory", MethodHandles.Lookup.class, String.class, MethodType.class, Object[].class);
 
         /*
          * Bootstrap method used to optimize String concatenation. Executing it at run time
@@ -141,6 +143,10 @@ public class BootstrapMethodConfiguration implements InternalFeature {
      */
     public boolean isIndyAllowedAtBuildTime(Executable method) {
         return method != null && indyBuildTimeAllowList.contains(method);
+    }
+
+    public boolean isMetafactory(Executable method) {
+        return method != null && (method.equals(metafactory) || method.equals(altMetafactory));
     }
 
     /**
