+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/heap/SVMImageLayerLoader.java
@@ -63,6 +63,7 @@ import com.oracle.graal.pointsto.heap.ImageLayerLoader;
 import com.oracle.graal.pointsto.meta.AnalysisMethod;
 import com.oracle.graal.pointsto.meta.AnalysisType;
 import com.oracle.graal.pointsto.util.AnalysisError;
+import com.oracle.graal.pointsto.util.AnalysisFuture;
 import com.oracle.svm.core.SubstrateOptions;
 import com.oracle.svm.core.TypeResult;
 import com.oracle.svm.core.classinitialization.ClassInitializationInfo;
@@ -208,10 +209,15 @@ public class SVMImageLayerLoader extends ImageLayerLoader {
     @Override
     protected boolean delegateProcessing(String constantType, Object constantValue, List<Object> constantData, Object[] values, int i) {
         if (constantType.equals(METHOD_POINTER_TAG)) {
-            AnalysisType methodPointerType = metaAccess.lookupJavaType(MethodPointer.class);
-            int mid = (int) constantValue;
-            AnalysisMethod method = getAnalysisMethod(mid);
-            values[i] = new RelocatableConstant(new MethodPointer(method), methodPointerType);
+            AnalysisFuture<JavaConstant> task = new AnalysisFuture<>(() -> {
+                AnalysisType methodPointerType = metaAccess.lookupJavaType(MethodPointer.class);
+                int mid = (int) constantValue;
+                AnalysisMethod method = getAnalysisMethod(mid);
+                RelocatableConstant constant = new RelocatableConstant(new MethodPointer(method), methodPointerType);
+                values[i] = constant;
+                return constant;
+            });
+            values[i] = task;
             return true;
         } else if (constantType.equals(C_ENTRY_POINT_LITERAL_CODE_POINTER)) {
             AnalysisType cEntryPointerLiteralPointerType = metaAccess.lookupJavaType(CEntryPointLiteralCodePointer.class);
